requirements.txt
dj-database-url>=0.4.2
dj-static>=0.0.6
django-taggit>=0.22
Django>=1.10
gunicorn>=19.7
pytz>=2017.2

